refactor: export interfaces, adapters, types and the lib
