+++ packages/email-auth/src/adapters/index.ts
@@ -1,0 +1,2 @@
+export { MongooseAdapter } from './mongoose-adapter';
+export { SendGridAdapter } from './send-grid-adapter';
