+++ packages/email-auth/src/interfaces/index.ts
@@ -1,0 +1,2 @@
+export { EmailableProviderAdapter } from './emailable-provider-adapter';
+export { PersistenceAdapter } from './persistence-adapter';
