+++ packages/email-auth/src/types/index.ts
